% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rags2ridgesFused.R
\name{GGMpathStats.fused}
\alias{GGMpathStats.fused}
\title{Fused gaussian graphical model node pair path statistics}
\usage{
GGMpathStats.fused(sparsePlist, ...)
}
\arguments{
\item{sparsePlist}{A \code{list} of sparsified precision matrices.}

\item{\dots}{Arguments passed to \code{\link{GGMpathStats}}.}
}
\value{
A \code{list} of path stats.
}
\description{
A simple wrapper for \code{\link{GGMpathStats}}.
}
\note{
The function currently fails if no paths are present in one of the
  groups.
}
\examples{
## Obtain some (high-dimensional) data
set.seed(1)
ns <- c(10, 11)
Slist <- createS(ns, p = 7, topology = "banded")
Tlist <- default.target.fused(Slist, ns)

## Obtain regularized precision and sparsify
Plist <- ridgeP.fused(Slist, ns, Tlist, lambda = c(1, 1.6))
sparsePlist <- sparsify.fused(Plist, threshold = "absValue", absValueCut = 0.20)
SPlist <- lapply(sparsePlist, "[[", "sparsePrecision")

## Obtain information on mediating and moderating paths between nodes 14 and 23
res <- GGMpathStats.fused(SPlist, node1 = 3, node2 = 4, graph = FALSE)

}
\seealso{
\code{\link{GGMpathStats}}
}
\author{
Anders E. Bilgrau, Carel F.W. Peeters <carel.peeters@wur.nl>, Wessel
  N. van Wieringen
}
